refactor: move webpack to root folder and rewrite to es6 node style (legacy compatibility)
